feat: Add Device entity with serial number, secret key, and owner ID
