+++ docs/session-management-design.md
@@ -84,6 +84,9 @@ deactivate SERVER
   "created_at": "Timestamp",
   "last_active_at": "Timestamp",
   "status": "active|inactive|revoked",
+  "serial_number": "Device serial number",
+  "secret_key": "Device secret key",
+  "owner_id": "Owner identifier",
   "metadata": {
     "model": "Device model",
     "os_version": "Operating system version",
