+++ server/domain/entities/entities.go
@@ -5,16 +5,6 @@ import (
 	"time"
 )
 
-// Device represents a doll device
-type Device struct {
-	ID           string    `json:"id" db:"id"`
-	SerialNumber string    `json:"serial_number" db:"serial_number"`
-	Model        string    `json:"model" db:"model"`
-	OwnerID      *string   `json:"owner_id" db:"owner_id"`
-	CreatedAt    time.Time `json:"created_at" db:"created_at"`
-	UpdatedAt    time.Time `json:"updated_at" db:"updated_at"`
-}
-
 // User represents a parent/user account
 type User struct {
 	ID        string    `json:"id" db:"id"`
